fix(e2e-tests): updates datepicker e2e tests to be more reliable
