+++ src/datepicker/__tests__/calendar.e2e.js
@@ -1,0 +1,24 @@
+/*
+Copyright (c) 2018-2019 Uber Technologies, Inc.
+
+This source code is licensed under the MIT license found in the
+LICENSE file in the root directory of this source tree.
+*/
+
+/* eslint-env node */
+/* eslint-disable flowtype/require-valid-file-annotation */
+
+const {mount, analyzeAccessibility} = require('../../../e2e/helpers');
+
+const selectors = {
+  calendar: '[role="application"]',
+};
+
+describe('Calendar', () => {
+  it('calendar passes basic a11y tests', async () => {
+    await mount(page, 'calendar');
+    await page.waitFor(selectors.calendar);
+    const accessibilityReport = await analyzeAccessibility(page);
+    expect(accessibilityReport).toHaveNoAccessibilityIssues();
+  });
+});
