+++ src/datepicker/__tests__/datepicker-int.e2e.js
@@ -1,0 +1,46 @@
+/*
+Copyright (c) 2018-2019 Uber Technologies, Inc.
+
+This source code is licensed under the MIT license found in the
+LICENSE file in the root directory of this source tree.
+*/
+
+/* eslint-env node */
+/* eslint-disable flowtype/require-valid-file-annotation */
+
+const {mount, analyzeAccessibility} = require('../../../e2e/helpers');
+
+const selectors = {
+  input: 'input',
+  calendar: '[role="application"]',
+  day: '[aria-label="Choose Sunday, March 10th 2019. It\'s available."]',
+  day2: '[aria-label="Choose Thursday, March 28th 2019. It\'s available."]',
+  monthYearSelectButton: '[data-id="monthYearSelectButton"]',
+  monthYearSelectMenu: '[data-id="monthYearSelectMenu"]',
+};
+
+describe('Datepicker, Int', () => {
+  it('selects range - int', async () => {
+    await mount(page, 'datepicker-int');
+    await page.waitFor(selectors.input);
+    await page.click(selectors.input);
+    await page.waitFor(selectors.calendar);
+    await page.click(selectors.day);
+    await page.waitFor(selectors.calendar);
+    const selectedValue1 = await page.$eval(
+      selectors.input,
+      input => input.value,
+    );
+    expect(selectedValue1).toBe('10.03.2019 -   .  .    ');
+
+    await page.click(selectors.day2);
+    await page.waitFor(selectors.calendar, {
+      hidden: true,
+    });
+    const selectedValue2 = await page.$eval(
+      selectors.input,
+      input => input.value,
+    );
+    expect(selectedValue2).toBe('10.03.2019 - 28.03.2019');
+  });
+});
