+++ src/datepicker/__tests__/datepicker-range.e2e.js
@@ -1,0 +1,96 @@
+/*
+Copyright (c) 2018-2019 Uber Technologies, Inc.
+
+This source code is licensed under the MIT license found in the
+LICENSE file in the root directory of this source tree.
+*/
+
+/* eslint-env node */
+/* eslint-disable flowtype/require-valid-file-annotation */
+
+const {mount, analyzeAccessibility} = require('../../../e2e/helpers');
+
+const selectors = {
+  input: 'input',
+  calendar: '[role="application"]',
+  day: '[aria-label="Choose Sunday, March 10th 2019. It\'s available."]',
+  day2: '[aria-label="Choose Thursday, March 28th 2019. It\'s available."]',
+  day4: '[aria-label="Choose Monday, April 1st 2019. It\'s available."]',
+  rightArrow: '[aria-label="Next month"]',
+  monthYearSelectButton: '[data-id="monthYearSelectButton"]',
+  monthYearSelectMenu: '[data-id="monthYearSelectMenu"]',
+};
+
+describe('Datepicker, Range', () => {
+  it('selects range - verifies end of year', async () => {
+    await mount(page, 'datepicker-range');
+
+    await page.waitFor(selectors.input);
+    await page.click(selectors.input);
+    await page.waitFor(selectors.calendar);
+    await page.click(selectors.monthYearSelectButton);
+    await page.waitFor(selectors.monthYearSelectMenu);
+
+    await page.$$eval('ul[role="listbox"] li', items => {
+      const option = items.find(item => {
+        return item.textContent === 'December 2019';
+      });
+      option.click();
+      return option;
+    });
+
+    await page.click(
+      '[aria-label="Choose Wednesday, December 25th 2019. It\'s available."]',
+    );
+
+    await page.click(
+      '[aria-label="Choose Tuesday, December 31st 2019. It\'s available."]',
+    );
+
+    const selectedValue = await page.$eval(
+      selectors.input,
+      input => input.value,
+    );
+    expect(selectedValue).toBe('2019/12/25 - 2019/12/31');
+  });
+
+  it('does not highlight random days when the right arrow is clicked', async () => {
+    await mount(page, 'datepicker-range');
+    await page.waitFor(selectors.input);
+    await page.click(selectors.input);
+    await page.waitFor(selectors.calendar);
+    await page.click(selectors.rightArrow);
+    await page.waitFor(selectors.day4);
+    const value = await page.$(selectors.day4);
+    const highlighted = await page.evaluate(
+      element => element.getAttribute('data-highlighted'),
+      value,
+    );
+    // Check the value of a custom data-highlighted attribute
+    expect(highlighted).toBe('false');
+  });
+
+  it('selects range', async () => {
+    await mount(page, 'datepicker-range');
+    await page.waitFor(selectors.input);
+    await page.click(selectors.input);
+    await page.waitFor(selectors.calendar);
+    await page.click(selectors.day);
+    await page.waitFor(selectors.calendar);
+    const selectedValue1 = await page.$eval(
+      selectors.input,
+      input => input.value,
+    );
+    expect(selectedValue1).toBe('2019/03/10 -     /  /  ');
+
+    await page.click(selectors.day2);
+    await page.waitFor(selectors.calendar, {
+      hidden: true,
+    });
+    const selectedValue2 = await page.$eval(
+      selectors.input,
+      input => input.value,
+    );
+    expect(selectedValue2).toBe('2019/03/10 - 2019/03/28');
+  });
+});
