+++ src/datepicker/__tests__/datepicker.e2e.js
@@ -26,13 +26,6 @@ const selectors = {
 };
 
 describe('Datepicker', () => {
-  it('calendar passes basic a11y tests', async () => {
-    await mount(page, 'calendar');
-    await page.waitFor(selectors.calendar);
-    const accessibilityReport = await analyzeAccessibility(page);
-    expect(accessibilityReport).toHaveNoAccessibilityIssues();
-  });
-
   it('datepicker passes basic a11y tests', async () => {
     await mount(page, 'datepicker');
     await page.waitFor(selectors.input);
@@ -111,170 +104,6 @@ describe('Datepicker', () => {
     expect(selectedValue).toBe('2019/03/18');
   });
 
-  it('selects range', async () => {
-    await mount(page, 'datepicker-range');
-    await page.waitFor(selectors.input);
-    await page.click(selectors.input);
-    await page.waitFor(selectors.calendar);
-    await page.click(selectors.day);
-    await page.waitFor(selectors.calendar);
-    const selectedValue1 = await page.$eval(
-      selectors.input,
-      input => input.value,
-    );
-    expect(selectedValue1).toBe('2019/03/10 -     /  /  ');
-
-    await page.click(selectors.day2);
-    await page.waitFor(selectors.calendar, {
-      hidden: true,
-    });
-    const selectedValue2 = await page.$eval(
-      selectors.input,
-      input => input.value,
-    );
-    expect(selectedValue2).toBe('2019/03/10 - 2019/03/28');
-  });
-
-  it('selects range - int', async () => {
-    await mount(page, 'datepicker-int');
-    await page.waitFor(selectors.input);
-    await page.click(selectors.input);
-    await page.waitFor(selectors.calendar);
-    await page.click(selectors.day);
-    await page.waitFor(selectors.calendar);
-    const selectedValue1 = await page.$eval(
-      selectors.input,
-      input => input.value,
-    );
-    expect(selectedValue1).toBe('10.03.2019 -   .  .    ');
-
-    await page.click(selectors.day2);
-    await page.waitFor(selectors.calendar, {
-      hidden: true,
-    });
-    const selectedValue2 = await page.$eval(
-      selectors.input,
-      input => input.value,
-    );
-    expect(selectedValue2).toBe('10.03.2019 - 28.03.2019');
-  });
-
-  it('selects range - verifies end of year', async () => {
-    await mount(page, 'datepicker-range');
-
-    await page.waitFor(selectors.input);
-    await page.click(selectors.input);
-    await page.waitFor(selectors.calendar);
-    await page.click(selectors.monthYearSelectButton);
-    await page.waitFor(selectors.monthYearSelectMenu);
-
-    await page.$$eval('ul[role="listbox"] li', items => {
-      const option = items.find(item => {
-        return item.textContent === 'December 2019';
-      });
-      option.click();
-      return option;
-    });
-
-    await page.click(
-      '[aria-label="Choose Wednesday, December 25th 2019. It\'s available."]',
-    );
-
-    await page.click(
-      '[aria-label="Choose Tuesday, December 31st 2019. It\'s available."]',
-    );
-
-    const selectedValue = await page.$eval(
-      selectors.input,
-      input => input.value,
-    );
-    expect(selectedValue).toBe('2019/12/25 - 2019/12/31');
-  });
-
-  it('renders previous month when the left arrow is clicked', async () => {
-    await mount(page, 'datepicker');
-    await page.waitFor(selectors.input);
-    await page.click(selectors.input);
-    await page.waitFor(selectors.calendar);
-    await page.click(selectors.leftArrow);
-    await page.waitFor(selectors.day, {
-      hidden: true,
-    });
-    await page.waitFor(selectors.day3);
-  });
-
-  it('renders next month when the right arrow is clicked', async () => {
-    await mount(page, 'datepicker');
-    await page.waitFor(selectors.input);
-    await page.click(selectors.input);
-    await page.waitFor(selectors.calendar);
-    await page.click(selectors.rightArrow);
-    await page.waitFor(selectors.day, {
-      hidden: true,
-    });
-    await page.waitFor(selectors.day4);
-  });
-
-  it('does not highlight random days when the right arrow is clicked', async () => {
-    await mount(page, 'datepicker-range');
-    await page.waitFor(selectors.input);
-    await page.click(selectors.input);
-    await page.waitFor(selectors.calendar);
-    await page.click(selectors.rightArrow);
-    await page.waitFor(selectors.day4);
-    const value = await page.$(selectors.day4);
-    const highlighted = await page.evaluate(
-      element => element.getAttribute('data-highlighted'),
-      value,
-    );
-    // Check the value of a custom data-highlighted attribute
-    expect(highlighted).toBe('false');
-  });
-
-  it('updates the calendar when a year selected from the dropdown', async () => {
-    await mount(page, 'datepicker');
-    await page.waitFor(selectors.input);
-    await page.click(selectors.input);
-    await page.waitFor(selectors.calendar);
-    await page.waitFor(selectors.day);
-    await page.click(selectors.monthYearSelectButton);
-    await page.waitFor(selectors.monthYearSelectMenu);
-
-    await page.$$eval('ul[role="listbox"] li', items => {
-      const option = items.find(item => {
-        return item.textContent === 'March 2018';
-      });
-      option.click();
-      return option;
-    });
-
-    await page.waitFor(selectors.monthYearSelectMenu, {hidden: true});
-    await page.waitFor(selectors.calendar);
-    await page.waitFor(selectors.day5);
-  });
-
-  it('updates the calendar when a month selected from the dropdown', async () => {
-    await mount(page, 'datepicker');
-    await page.waitFor(selectors.input);
-    await page.click(selectors.input);
-    await page.waitFor(selectors.calendar);
-    await page.waitFor(selectors.day);
-    await page.click(selectors.monthYearSelectButton);
-    await page.waitFor(selectors.monthYearSelectMenu);
-
-    await page.$$eval('ul[role="listbox"] li', items => {
-      const option = items.find(item => {
-        return item.textContent === 'July 2019';
-      });
-      option.click();
-      return option;
-    });
-
-    await page.waitFor(selectors.monthYearSelectMenu, {hidden: true});
-    await page.waitFor(selectors.calendar);
-    await page.waitFor(selectors.day6);
-  });
-
   it('month year dropdown opens on arrow down', async () => {
     await mount(page, 'datepicker');
     await page.waitFor(selectors.input);
@@ -310,53 +139,5 @@ describe('Datepicker', () => {
     await page.waitFor(selectors.monthYearSelectMenu);
     await page.keyboard.press('Tab');
     await page.waitFor(selectors.monthYearSelectMenu, {hidden: true});
-  });
-
-  it('disables previous month button if minimum month is selected', async () => {
-    await mount(page, 'datepicker');
-    await page.waitFor(selectors.input);
-    await page.click(selectors.input);
-    await page.waitFor(selectors.calendar);
-    await page.waitFor(selectors.day);
-    await page.click(selectors.monthYearSelectButton);
-    await page.waitFor(selectors.monthYearSelectMenu);
-
-    await page.$$eval('ul[role="listbox"] li', items => {
-      const option = items.find(item => {
-        return item.textContent === 'January 2000';
-      });
-      option.click();
-      return option;
-    });
-
-    await page.click(selectors.leftArrow);
-    const value = await page.$(selectors.monthYearSelectButton);
-    const text = await page.evaluate(element => element.textContent, value);
-    // (Month YearTriangle Down) because it renders an icon within the element
-    expect(text).toBe('January 2000Triangle Down');
-  });
-
-  it('disables next month button if maximum month is selected', async () => {
-    await mount(page, 'datepicker');
-    await page.waitFor(selectors.input);
-    await page.click(selectors.input);
-    await page.waitFor(selectors.calendar);
-    await page.waitFor(selectors.day);
-    await page.click(selectors.monthYearSelectButton);
-    await page.waitFor(selectors.monthYearSelectMenu);
-
-    await page.$$eval('ul[role="listbox"] li', items => {
-      const option = items.find(item => {
-        return item.textContent === 'December 2030';
-      });
-      option.click();
-      return option;
-    });
-
-    await page.click(selectors.rightArrow);
-    const value = await page.$(selectors.monthYearSelectButton);
-    const text = await page.evaluate(element => element.textContent, value);
-    // (Month YearTriangle Down) because it renders an icon within the element
-    expect(text).toBe('December 2030Triangle Down');
   });
 });
